Renaming library to resource_abstractor_client
