+++ libraries/resource_abstractor_client/client_helper.py
