+++ libraries/resource_abstractor_client/cluster_operations.py
@@ -1,5 +1,5 @@
-from rasclient.client_helper import make_request
 from requests import get, patch, put
+from resource_abstractor_client.client_helper import make_request
 
 RESOURCES_API = "/api/v1/resources"
