+++ libraries/setup.py
@@ -1,7 +1,7 @@
 from setuptools import find_packages, setup
 
 setup(
-    name="rasclient",
+    name="resource_abstractor_client",
     version="0.1.0",
     packages=find_packages(),
     include_package_data=True,
